refactor(subscriptions): move strip_null_bytes to delivery_common
Address PR review:
- Move strip_null_bytes to delivery_common (a general-purpose sanitizer with no
  dependency on insight types); re-export from insight_snapshot so existing
  importers keep resolving.
- Handle tuples in strip_null_bytes for parity with lists, so a future caller
  passing a tuple (e.g. an un-flattened diagnostics collection) still gets
  scrubbed instead of silently passing NUL through.
- Add AI-report test coverage with NUL-bearing markdown, diagnostics, and prompt
  so a dropped scrub is caught.

Generated-By: PostHog Code
Task-Id: 900469b2-6987-469a-a26b-3f040acd912f
